My Thoughts
For this project I was coding in Python, and the goal was to create a program that would output a random roll of a 6 sided die. The program begins by asking the user if they would like to roll. Supposing the answer is no, the program ends. If the answer is yes, the program will continue to ask infinitely, or until the user answers no. The hardest part of this was accounting for something other than a "yes" or "no" input. I would like to know if there is a way to make this same program with less code. I could not seem to find a way to merge if statements and while statements in such a way that it could fit into one block. As it stands, if the first input is a yes, that input and all subsequent "yes's" will execute within the first while loop. If the user inputs anything other than yes or no at any time, they are brought into the second while loop, which accounts for both "yes" inputs and improper inputs infinitely. And if at any time they answer no in either while loop, the corresponding if statement executes.